Group the MedPC edge cases beside the on-data tests
